requirements.txt
pandas
numpy
matplotlib
statsmodels
scikit-learn
tensorflow
streamlit
pyyaml
plotly
